Laravel: hook restructuring.
