+++ src/Instrumentation/Laravel/src/Hooks/Illuminate/Queue/Worker.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Contrib\Instrumentation\Laravel\Hooks\Queue;
+namespace OpenTelemetry\Contrib\Instrumentation\Laravel\Hooks\Illuminate\Queue;
 
 use Illuminate\Contracts\Queue\Job;
 use Illuminate\Queue\Worker as QueueWorker;
